requirements.txt
paddlepaddle-gpu>=2.4.0
paddlenlp>=2.5.0
numpy>=1.21.0
tqdm>=4.64.0
sumeval>=0.6.1